# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface WeatherLocation {
  name: string;
  /**
     * @minimum 24
     * @maximum 50
     */
  latitude: number;
  /**
     * @minimum -125
     * @maximum -66
     */
  longitude: number;
}

export type ChatTurnRole = typeof ChatTurnRole[keyof typeof ChatTurnRole];


export const ChatTurnRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatTurn {
  role: ChatTurnRole;
  /** @maxLength 6000 */
  content: string;
}

export interface WeatherChatInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  location?: WeatherLocation;
  /** @maxItems 12 */
  history?: ChatTurn[];
}

export interface WeatherChatResult {
  answer: string;
  location?: WeatherLocation;
  sources: string[];
}

export interface HealthStatus {
  status: string;
}