# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /weather/chat:
    post:
      operationId: weatherChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeatherChatInput"
      responses:
        "200":
          description: Grounded weather answer and optional map destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherChatResult"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    WeatherChatInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1, maxLength: 2000 }
        location:
          $ref: "#/components/schemas/WeatherLocation"
        history:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/ChatTurn"
    ChatTurn:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string, maxLength: 6000 }
    WeatherLocation:
      type: object
      required: [name, latitude, longitude]
      properties:
        name: { type: string }
        latitude: { type: number, minimum: 24, maximum: 50 }
        longitude: { type: number, minimum: -125, maximum: -66 }
    WeatherChatResult:
      type: object
      required: [answer, sources]
      properties:
        answer: { type: string }
        location:
          $ref: "#/components/schemas/WeatherLocation"
        sources:
          type: array
          items: { type: string }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status