# lib/api-zod/src/generated/types/weatherChatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatTurn } from './chatTurn';
import type { WeatherLocation } from './weatherLocation';

export interface WeatherChatInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  location?: WeatherLocation;
  /** @maxItems 12 */
  history?: ChatTurn[];
}