# lib/api-zod/src/generated/types/weatherChatResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WeatherLocation } from './weatherLocation';

export interface WeatherChatResult {
  answer: string;
  location?: WeatherLocation;
  sources: string[];
}