# lib/api-zod/src/generated/types/weatherLocation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherLocation {
  name: string;
  /**
     * @minimum 24
     * @maximum 50
     */
  latitude: number;
  /**
     * @minimum -125
     * @maximum -66
     */
  longitude: number;
}